+++ R/zzz.R
@@ -1,7 +1,0 @@
-.onAttach <- function(...) {
-
-packageStartupMessage(
-"From now on difference in fariness_check and parity_loss
-has been changed to ratios to adhere to four-fifths (80%) rule (adverse impact).
-More on that in documentation. Thank you for using fairmodels!\n")
-}